Prepare routes to add svg prop animations examples
