+++ apps/common-app/src/apps/css/examples/animations/routes/properties/svg.ts
@@ -1,0 +1,5 @@
+import type { Routes } from '@/apps/css/navigation/types';
+
+export const svgAnimatedPropertiesRoutes = {
+  // TODO - add svg example routes here
+} satisfies Routes;
